Move check_from_class to ballet.validation.common
